requirements.txt
django>=3.2.14

Pillow>=9.0
